requirements.txt
streamlit
pyspf
pandas